refactor(api): extract read layer into pkg/api/internal/dbreads
Move every DB-read method off *api.API into a new pkg/api/internal/dbreads.Queries
type, which *API embeds. Reads are promoted onto *API so all external
callers (pkg/handler, pkg/omaha, pkg/syncer, pkg/metrics, pkg/server,
test/*) compile unchanged. Write logic stays on *API.

This is roadmap PR 3 from RFC #1375. It unblocks the follow-up split
of admin vs runtime writer packages by isolating the read surface both
will need to embed.

Refs: #1375
