+++ backend/pkg/api/groups_test.go
@@ -36,7 +36,7 @@ func TestAddGroup(t *testing.T) {
 	assert.NoError(t, err)
 	assert.Equal(t, true, group.PolicyUpdatesEnabled)
 
-	_, err = a.getGroupUpdatesStats(group)
+	_, err = a.GetGroupUpdatesStats(group)
 	assert.NoError(t, err)
 
 	groupX, err := a.GetGroup(group.ID)
@@ -76,7 +76,7 @@ func TestUpdateGroup(t *testing.T) {
 	err := a.UpdateGroup(group)
 	assert.NoError(t, err)
 
-	_, err = a.getGroupUpdatesStats(group)
+	_, err = a.GetGroupUpdatesStats(group)
 	assert.NoError(t, err)
 
 	groupX, _ := a.GetGroup(group.ID)
