+++ backend/pkg/api/helpers.go
@@ -2,28 +2,7 @@ package api
 
 import (
 	"time"
-
-	"github.com/blang/semver/v4"
 )
-
-const (
-	defaultPage    uint64 = 1
-	defaultPerPage uint64 = 10
-)
-
-// validatePaginationParams validates the pagination parameters provided,
-// setting them to the default values in case they are invalid.
-func validatePaginationParams(page, perPage uint64) (uint64, uint64) {
-	if page < 1 {
-		page = defaultPage
-	}
-
-	if perPage < 1 {
-		perPage = defaultPerPage
-	}
-
-	return page, perPage
-}
 
 // isTimezoneValid checks if the provided timezone is valid.
 func isTimezoneValid(tz string) bool {
@@ -36,17 +15,4 @@ func isTimezoneValid(tz string) bool {
 	}
 
 	return true
-}
-
-// isValidSemver checks if the provided string represents a valid semver
-// version.
-func isValidSemver(version string) bool {
-	if _, err := semver.Make(version); err != nil {
-		return false
-	}
-	return true
-}
-
-func sqlPaginate(page, perPage uint64) (uint, uint) {
-	return uint(perPage), uint(page-1) * uint(perPage)
 }
