+++ backend/pkg/api/internal/dbreads/helpers.go
@@ -1,0 +1,35 @@
+package dbreads
+
+import "github.com/blang/semver/v4"
+
+const (
+	defaultPage    uint64 = 1
+	defaultPerPage uint64 = 10
+)
+
+// validatePaginationParams validates the pagination parameters provided,
+// setting them to the default values in case they are invalid.
+func validatePaginationParams(page, perPage uint64) (uint64, uint64) {
+	if page < 1 {
+		page = defaultPage
+	}
+
+	if perPage < 1 {
+		perPage = defaultPerPage
+	}
+
+	return page, perPage
+}
+
+// IsValidSemver checks if the provided string represents a valid semver
+// version.
+func IsValidSemver(version string) bool {
+	if _, err := semver.Make(version); err != nil {
+		return false
+	}
+	return true
+}
+
+func sqlPaginate(page, perPage uint64) (uint, uint) {
+	return uint(perPage), uint(page-1) * uint(perPage)
+}
