+++ backend/pkg/api/packages.go
@@ -4,13 +4,12 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
-	"strings"
 
 	"github.com/doug-martin/goqu/v9"
 	"github.com/google/uuid"
 	"github.com/jmoiron/sqlx"
-	"gopkg.in/guregu/null.v4"
 
+	"github.com/flatcar/nebraska/backend/pkg/api/internal/dbreads"
 	"github.com/flatcar/nebraska/backend/pkg/api/internal/types"
 )
 
@@ -80,7 +79,7 @@ func (api *API) checkMatchingArch(channelIDs StringArray, arch Arch) error {
 // It handles validation, package insertion, blacklist, and files within a transaction.
 // The caller is responsible for handling FlatcarAction and committing the transaction.
 func (api *API) addPackage(pkg *Package, tx *sqlx.Tx) error {
-	if !isValidSemver(pkg.Version) {
+	if !dbreads.IsValidSemver(pkg.Version) {
 		return ErrInvalidSemver
 	}
 	if !pkg.Arch.IsValid() {
@@ -239,7 +238,7 @@ func (api *API) AddPackageWithMetadata(pkg *Package) (*Package, error) {
 // UpdatePackage updates an existing package using the content of the package
 // provided.
 func (api *API) UpdatePackage(pkg *Package) error {
-	if !isValidSemver(pkg.Version) {
+	if !dbreads.IsValidSemver(pkg.Version) {
 		return ErrInvalidSemver
 	}
 	tx, err := api.db.Beginx()
@@ -365,266 +364,6 @@ func (api *API) DeletePackage(pkgID string) error {
 	return nil
 }
 
-// GetPackage returns the package identified by the id provided.
-func (api *API) GetPackage(pkgID string) (*Package, error) {
-	return api.getPackage(null.StringFrom(pkgID))
-}
-
-// GetPackageByVersionAndArch returns the package identified by the
-// application ID, version and arch provided.
-func (api *API) GetPackageByVersionAndArch(appID, version string, arch Arch) (*Package, error) {
-	var pkg Package
-	if !isValidSemver(version) {
-		return nil, fmt.Errorf("error GetPackageByVersionAndArch version %s is not valid", version)
-	}
-	query, _, err := api.packagesQuery().
-		Where(goqu.C("application_id").Eq(appID), goqu.C("arch").Eq(arch), goqu.C("version").Eq(version)).
-		ToSQL()
-	if err != nil {
-		return nil, err
-	}
-	err = api.db.QueryRowx(query).StructScan(&pkg)
-	if err != nil {
-		return nil, err
-	}
-	flatcarAction, err := api.getFlatcarAction(pkg.ID)
-	switch err {
-	case nil:
-		pkg.FlatcarAction = flatcarAction
-	case sql.ErrNoRows:
-		pkg.FlatcarAction = nil
-	default:
-		return nil, err
-	}
-	return &pkg, nil
-}
-
-// GetPackagesCount retuns the total number of package in an app
-func (api *API) GetPackagesCount(appID string, searchVersion *string) (int, error) {
-	query := goqu.From(goqu.L("package LEFT JOIN package_channel_blacklist pcb ON package.id = pcb.package_id")).
-		Select(goqu.L(`package.*,
-	    array_agg(pcb.channel_id) FILTER (WHERE pcb.channel_id IS NOT NULL) as channels_blacklist
-	    `)).Where(goqu.C("application_id").Eq(appID)).
-		GroupBy("package.id")
-	if searchVersion != nil {
-		*searchVersion = "%" + strings.ToLower(*searchVersion) + "%"
-		query = query.Where(goqu.I("version").ILike(*searchVersion))
-	}
-	query = goqu.From(query).Select(goqu.L("count (*)"))
-	return api.GetCountQuery(query)
-}
-
-// GetPackages returns all packages associated to the application provided.
-func (api *API) GetPackages(appID string, page, perPage uint64, searchVersion *string) ([]*Package, error) {
-	page, perPage = validatePaginationParams(page, perPage)
-	limit, offset := sqlPaginate(page, perPage)
-	query := api.packagesQuery().
-		Where(goqu.C("application_id").Eq(appID)).
-		Limit(limit).
-		Offset(offset)
-	if searchVersion != nil {
-		*searchVersion = "%" + strings.ToLower(*searchVersion) + "%"
-		query = query.Where(goqu.I("version").ILike(*searchVersion))
-	}
-	queryString, _, err := query.ToSQL()
-	if err != nil {
-		return nil, err
-	}
-
-	return api.getPackagesFromQuery(queryString)
-}
-
-func (api *API) getPackagesFromQuery(query string) ([]*Package, error) {
-	var pkgs []*Package
-	rows, err := api.db.Queryx(query)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	// Load all packages
-	for rows.Next() {
-		pkg := Package{}
-		err = rows.StructScan(&pkg)
-		if err != nil {
-			return nil, err
-		}
-		pkgs = append(pkgs, &pkg)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, err
-	}
-
-	if len(pkgs) == 0 {
-		return pkgs, nil
-	}
-
-	// Use loadPackageExtras to batch load extra files and actions
-	return api.loadPackageExtras(pkgs)
-}
-
-// loadPackageExtras loads extra files and flatcar actions for packages efficiently
-func (api *API) loadPackageExtras(packages []*Package) ([]*Package, error) {
-	if len(packages) == 0 {
-		return packages, nil
-	}
-
-	// Collect package IDs
-	pkgIDs := make([]string, len(packages))
-	for i, pkg := range packages {
-		pkgIDs[i] = pkg.ID
-	}
-
-	// Load extra files
-	query, _, err := goqu.From("package_file").
-		Where(goqu.C("package_id").In(pkgIDs)).
-		Order(goqu.C("package_id").Asc(), goqu.C("id").Asc()).
-		ToSQL()
-	if err != nil {
-		return nil, err
-	}
-
-	var files []File
-	if err := api.db.Select(&files, query); err != nil {
-		return nil, err
-	}
-
-	filesByPkg := make(map[string][]File)
-	for _, file := range files {
-		filesByPkg[file.PackageID] = append(filesByPkg[file.PackageID], file)
-	}
-
-	// Load Flatcar actions
-	query, _, err = goqu.From("flatcar_action").
-		Where(goqu.C("package_id").In(pkgIDs)).
-		ToSQL()
-	if err != nil {
-		return nil, err
-	}
-
-	var actions []FlatcarAction
-	if err := api.db.Select(&actions, query); err != nil && err != sql.ErrNoRows {
-		return nil, err
-	}
-
-	actionsByPkg := make(map[string]*FlatcarAction)
-	for i := range actions {
-		actionsByPkg[actions[i].PackageID] = &actions[i]
-	}
-
-	// Assign files and actions to packages
-	for _, pkg := range packages {
-		if files, ok := filesByPkg[pkg.ID]; ok {
-			pkg.ExtraFiles = files
-		} else {
-			pkg.ExtraFiles = nil
-		}
-		if action, ok := actionsByPkg[pkg.ID]; ok {
-			pkg.FlatcarAction = action
-		} else {
-			pkg.FlatcarAction = nil
-		}
-	}
-
-	return packages, nil
-}
-
-// packagesQuery returns a SelectDataset prepared to return all packages.
-// This query is meant to be extended later in the methods using it to filter
-// by a specific package id, all packages that belong to a given application,
-// specify how to query the rows or their destination.
-func (api *API) packagesQuery() *goqu.SelectDataset {
-	// Note: semverToIntArray error handling is deferred to when ToSQL() is called
-	// since goqu.SelectDataset doesn't support immediate error returns
-	semverExpr, err := semverToIntArray("version")
-	if err != nil {
-		// Return an invalid query that will fail when ToSQL() is called
-		return goqu.From("invalid_table_error_" + err.Error())
-	}
-
-	query := goqu.From(goqu.L("package LEFT JOIN package_channel_blacklist pcb ON package.id = pcb.package_id")).
-		Select(goqu.L(`package.*,
-	    array_agg(pcb.channel_id) FILTER (WHERE pcb.channel_id IS NOT NULL) as channels_blacklist
-	    `)).
-		GroupBy("package.id").Order(goqu.L(semverExpr).Desc())
-	return query
-}
-func (api *API) getFlatcarActionQuery(packageID string) *goqu.SelectDataset {
-	query := goqu.From("flatcar_action").Where(goqu.C("package_id").Eq(packageID))
-	return query
-}
-func (api *API) getFlatcarAction(packageID string) (*FlatcarAction, error) {
-	query, _, err := api.getFlatcarActionQuery(packageID).ToSQL()
-	if err != nil {
-		return nil, err
-	}
-	flatcarAction := FlatcarAction{}
-	err = api.db.QueryRowx(query).StructScan(&flatcarAction)
-	if err != nil {
-		return nil, err
-	}
-	return &flatcarAction, nil
-}
-
-func (api *API) getExtraFiles(packageID string) ([]File, error) {
-	query, _, err := goqu.From("package_file").Where(goqu.C("package_id").Eq(packageID)).Order(goqu.C("id").Asc()).ToSQL()
-	if err != nil {
-		return nil, err
-	}
-
-	var files []File
-	rows, err := api.db.Queryx(query)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	for rows.Next() {
-		f := File{}
-		if err := rows.StructScan(&f); err != nil {
-			return nil, err
-		}
-
-		files = append(files, f)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, err
-	}
-	return files, nil
-}
-
-func (api *API) getPackage(packageID null.String) (*Package, error) {
-	query, _, err := api.packagesQuery().Where(goqu.C("id").Eq(packageID)).ToSQL()
-	if err != nil {
-		return nil, err
-	}
-	packageEntity := Package{}
-	err = api.db.QueryRowx(query).StructScan(&packageEntity)
-	if err != nil {
-		return nil, err
-	}
-	flatcarAction, err := api.getFlatcarAction(packageEntity.ID)
-	switch err {
-	case nil:
-		packageEntity.FlatcarAction = flatcarAction
-	case sql.ErrNoRows:
-		packageEntity.FlatcarAction = nil
-	default:
-		return nil, err
-	}
-
-	extraFiles, err := api.getExtraFiles(packageEntity.ID)
-	switch err {
-	case nil:
-		packageEntity.ExtraFiles = extraFiles
-	case sql.ErrNoRows:
-		packageEntity.ExtraFiles = nil
-	default:
-		return nil, err
-	}
-
-	return &packageEntity, nil
-}
-
 // updatePackageBlacklistedChannels adds or removes as needed channels to the
 // package's channels blacklist based on the new entries provided in the updated
 // package entry.
@@ -657,7 +396,7 @@ func (api *API) updatePackageBlacklistedChannels(tx *sqlx.Tx, pkg *Package, oldP
 			return ErrBlacklistingChannel
 		}
 		// Check if package is a floor for this channel
-		isFloor, err := api.isPackageFloorForChannel(pkg.ID, channelID)
+		isFloor, err := api.IsPackageFloorForChannel(pkg.ID, channelID)
 		if err != nil {
 			return err
 		}
