+++ backend/pkg/api/users.go
@@ -41,50 +41,6 @@ func (api *API) AddUser(user *User) (*User, error) {
 	return user, nil
 }
 
-// GetUser returns the user identified by the username provided.
-func (api *API) GetUser(username string) (*User, error) {
-	var user User
-	query, _, err := goqu.From("users").
-		Where(goqu.C("username").Eq(username)).
-		ToSQL()
-	if err != nil {
-		return nil, err
-	}
-	err = api.db.QueryRowx(query).StructScan(&user)
-	if err != nil {
-		return nil, err
-	}
-
-	return &user, nil
-}
-
-func (api *API) GetUsersInTeam(teamID string) ([]*User, error) {
-	var users []*User
-	query, _, err := goqu.From("users").
-		Where(goqu.C("team_id").Eq(teamID)).
-		ToSQL()
-	if err != nil {
-		return nil, err
-	}
-	rows, err := api.db.Queryx(query)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var user User
-		err := rows.StructScan(&user)
-		if err != nil {
-			return nil, err
-		}
-		users = append(users, &user)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, err
-	}
-	return users, nil
-}
-
 // UpdateUserPassword updates the password of the provided user.
 func (api *API) UpdateUserPassword(username, newPassword string) error {
 	secret, err := api.GenerateUserSecret(username, newPassword)
